Time: 0 ms (100.00%), Space: 5.7 MB (88.79%) - LeetHub

+++ 1025-divisor-game/1025-divisor-game.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    bool divisorGame(int n) {
+    bool divisorGame(short n) {
         return !(n&1);
     }
 };
